+++ HISTORY.md
@@ -8,6 +8,7 @@
 * **Notable Changes:**
     * searching for blocks and keyboard entry now includes the contents of dropdown menus
     * disabled dropping reporters into certain dropdowns (monadic functions, types, costume attributees, graphic effects, layers, audio attributes, pen attributes, dates, relation, keys, video attributes)
+    * changed VIDEOP _ ON _ reporter primitive to be hyper-monadic (second slot)
 * **Notable Fixes:**
     * keep internal linked-list organization intact for hyperblocks
 
@@ -19,6 +20,7 @@
 * blocks: made audio attributes menu static
 * blocks: made pen attributes menus static
 * blocks: made sensing attributes menus largely static
+* threads: changed reportVideo() to be hyper-monadic
 
 ### 2020-11-30
 * threads: keep internal linked-list organization intact for hyperblocks